Update 28JAN 12:57AM

+++ src/main/java/com/rms/repositories/OrganizationRepository.java
@@ -2,10 +2,9 @@ package com.rms.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.rms.persistences.Organization;
 
 @Repository
-public interface OrganizationRepository extends JpaRepository<Organization, Integer>{
+public interface OrganizationRepository extends JpaRepository<Organization, Long>{
 	
 }
